adding tests for images and renaming files
