+++ autogoal/experimental/augly/__init__.py
@@ -1,4 +1,4 @@
-from ._base import (
+from ._text import (
     InsertPunctuation,
     InsertZeroWidth,
     ReplaceBidirectional,
@@ -8,3 +8,24 @@ from ._base import (
     SimulateTypos,
     SplitWords
 )
+
+from ._image import (
+    Brightness,
+    ChangeAspectRatio,
+    Contrast,
+    Crop,
+    Grayscale,
+    Opacity,
+    Pixelization,
+    Rotate,
+    Resize,
+    Saturation,
+)
+
+from ._semantics import (
+    Image as AugLyImage,
+)
+
+import warnings
+
+warnings.warn(" AugLy integration is an experimental feature that has not been properly tested, it is not recommended to use it in production ")
